requirements.txt
google-cloud-bigquery
google-cloud
google-api-core
google-cloud-pubsub